requirements.txt
numpy
torch
torchvision
easyocr
openvino-dev
opencv-python